method that belongs to castsValidatedData
